Add files via upload
This modifies the last 'faceVerify' code by revising the frame sample number for each identity. Through analyzing the faceVerify result by the last code,  it shows that the more of the sampling frames for a same person, and the less number of different identity, which will result higher P/N value(P represents the number of positive face pairs that correspond to a same identity, and N is the number of negative pairs that correspond to different persons), will produce better recall, precision, and finally better F1 and Accuracy.

But in the last version of face pair files, only 3 frames are selected for each person, and the number of different persons is large, so the above measurement parameters are small and bad. So, firstly, it regenerate face pair files in which there are 4 identities and  20 frames are selected for each of them. In this way, it tries to promote the recall, precision, and F1, Accuracy value. Then, secondly, the 'faceVerify' is revised according to this.

That's the reason of producing this code.